requirements.txt
asgiref==3.5.1
Django==4.0.4
django-cors-headers==3.12.0
django-countries==7.3.2
django-localflavor==3.1
django-phone-field==1.8.1
djangorestframework==3.13.1
isoweek==1.3.3
python-stdnum==1.17
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.2.0
